requirements.txt
fastapi
uvicorn[standard]
streamlit
pinecone-client
requests
python-multipart
pydantic
pdfplumber
python-docx
transformers
sentence-transformers
torch
accelerate
tqdm
huggingface_hub


transformers==4.42.4
huggingface_hub>=0.23.0
accelerate>=0.30.0
tokenizers>=0.15.2
